requirements.txt
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-debug==0.1.1
django-debug-toolbar-template-profiler==2.1.0
django-extensions==3.2.1
django-js-asset==2.0.0
django_debug_toolbar==3.8.1
executing==1.2.0
idna==3.4
ipython==8.10.0
ipython-genutils==0.2.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydotplus==2.0.2
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.16.2
pyparsing==3.0.9
python-decouple==3.7
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
wrapt==1.14.1
